+++ ui/src/Containers/VulnMgmt/Entity/Cve/VulnMgmtEntityCve.js
@@ -23,6 +23,7 @@ const VulmMgmtCve = ({ entityId, entityListType, search, entityContext, sort, pa
             result: vulnerability(id: $id) {
                 id: cve
                 cve
+                vulnerabilityType
                 envImpact
                 cvss
                 scoreVersion