+++ ui/src/Containers/VulnMgmt/VulnMgmt.fragments.js
@@ -60,6 +60,7 @@ export const VULN_CVE_LIST_FRAGMENT = gql`
         id: cve
         cve
         cvss
+        vulnerabilityType
         scoreVersion
         envImpact
         impactScore