+++ CHANGELOG.md
@@ -12,6 +12,8 @@ All notable changes to this project that require documentation updates will be d
   6. `image` option  will be renamed to `main-image-repository`
   7. `collector-image` option will be renamed to `collector-image-repository`
   8. `monitoring-endpoint` option, which has already been deprecated, will be removed
+- Add CVE Type to CVE list and overview pages (ROX-4482)
+
 ## [44.0]
 - Previously, a scan for an image that may have been retagged (e.g. using the latest tag) would return a stale scan if it had been previously scanned.
 - UI: In Platform Configuration > Interactions: 1. replace "AWS ECR" with "Amazon ECR" and 2. replace "S3" (and "AWS S3" placeholder for Integration Name in New Integration pane) with "Amazon S3" (ROX-4912)